Creation and implementation of "filename_validation.py" file in "utils" folder. Code cleaning.
